import soul storage model.
Took 2 hours 2 minutes

+++ src/main/java/com/github/x3rdev/soul_forge/common/block_entity/SoulStorageBlockEntity.java
@@ -64,8 +64,8 @@ public class SoulStorageBlockEntity extends BlockEntity implements GeoBlockEntit
         }
     }
 
-    public double getTick() {
-        return this.level.getGameTime() - tickLoaded;
+    public int getTick() {
+        return (int) (this.level.getGameTime() - tickLoaded);
     }
 
     @Override
